Rename attribute generation file
