+++ .github/workflows/python-package.yml
@@ -46,7 +46,7 @@ jobs:
         run: |
           pytest tests/phantom_wiki/core
           pytest tests/phantom_wiki/facts/family
-          pytest tests/phantom_wiki/facts/test_generate.py
+          pytest tests/phantom_wiki/facts/test_generate_attributes.py
           pytest tests/phantom_wiki/facts/test_get_names.py
           pytest tests/phantom_wiki/facts/test_load_database.py
           pytest tests/phantom_wiki/facts/test_save_database.py
