+++ src/phantom_wiki/facts/attributes/generate_attributes.py
